move splitter cli tests into own directory
write tests for obvious (ROI, nz) configurations
