+++ tests/modules/mesoscope_splitting_cli/conftest.py
@@ -3,51 +3,12 @@ import copy
 import tifffile
 import pathlib
 import h5py
-from unittest.mock import patch
 import tempfile
 import numpy as np
 from ophys_etl.utils.array_utils import normalize_array
-from ophys_etl.modules.mesoscope_splitting.__main__ import (
-    TiffSplitterCLI)
 
 
-@pytest.fixture(scope='session')
-def splitter_tmp_dir_fixture(tmp_path_factory):
-    tmp_dir = tmp_path_factory.mktemp('splitter_cli_test')
-    return pathlib.Path(tmp_dir)
-
-
-@pytest.fixture(scope='session')
-def z_to_exp_id_fixture():
-    """
-    dict mapping z value to ophys_experiment_id
-    """
-    return {(2, 1): {2: 222, 1: 111},
-            (3, 4): {3: 333, 4: 444},
-            (6, 5): {6: 666, 5: 555},
-            (8, 7): {8: 888, 7: 777}}
-
-
-@pytest.fixture(scope='session')
-def z_list_fixture():
-    z_list = [[2, 1], [3, 4], [6, 5], [8, 7]]
-    return z_list
-
-
-@pytest.fixture(scope='session')
-def roi_index_to_z_fixture():
-    return {0: [1, 2, 3, 4], 1: [5, 6, 7, 8]}
-
-
-@pytest.fixture(scope='session')
-def z_to_roi_index_fixture():
-    return {(2, 1): 0,
-            (3, 4): 0,
-            (6, 5): 1,
-            (8, 7): 1}
-
-
-@pytest.fixture(scope='session')
+@pytest.fixture
 def z_to_stack_path_fixture(splitter_tmp_dir_fixture,
                             z_list_fixture):
     tmp_dir = splitter_tmp_dir_fixture
@@ -60,7 +21,7 @@ def z_to_stack_path_fixture(splitter_tmp_dir_fixture,
     return result
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def image_metadata_fixture(z_list_fixture,
                            roi_index_to_z_fixture):
     """
@@ -85,12 +46,15 @@ def image_metadata_fixture(z_list_fixture,
                                    {'centerXY': [9+roi_index, 10-roi_index]}]}
         roi_list.append(this_roi)
 
+    if len(roi_list) == 1:
+        roi_list = roi_list[0]
+
     metadata.append({'RoiGroups': {'imagingRoiGroup': {'rois': roi_list}}})
 
     return metadata
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def surface_metadata_fixture(image_metadata_fixture,
                              roi_index_to_z_fixture):
     """
@@ -103,29 +67,37 @@ def surface_metadata_fixture(image_metadata_fixture,
     for ii in range(n_rois):
         val = 1+17*ii
         z_list.append([val, 0])
-        rois[ii]['zs'] = val
+        if n_rois > 1:
+            rois[ii]['zs'] = val
+        else:
+            rois['zs'] = val
 
     metadata[0]['SI.hStackManager.zsAllActuators'] = z_list
     return metadata
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def zstack_metadata_fixture(splitter_tmp_dir_fixture,
                             image_metadata_fixture,
                             z_to_stack_path_fixture,
-                            z_to_roi_index_fixture):
+                            z_to_roi_index_fixture,
+                            roi_index_to_z_fixture):
     """
     Dict mapping z-stack path to metadata for the z-stack file
     """
+    n_rois = len(roi_index_to_z_fixture)
     result = dict()
     for pair in z_to_stack_path_fixture:
         this_path = z_to_stack_path_fixture[pair]
         this_metadata = copy.deepcopy(image_metadata_fixture)
         rois = this_metadata[1]['RoiGroups']['imagingRoiGroup']['rois']
-        for ii in range(len(rois)):
-            rois[ii]['discretePlaneMode'] = 1
-        roi_index = z_to_roi_index_fixture[tuple(pair)]
-        rois[roi_index]['discretePlaneMode'] = 0
+        if n_rois > 1:
+            for ii in range(len(rois)):
+                rois[ii]['discretePlaneMode'] = 1
+            roi_index = z_to_roi_index_fixture[tuple(pair)]
+            rois[roi_index]['discretePlaneMode'] = 0
+        else:
+            rois['discretePlaneMode'] = 0
         z0_list = np.linspace(pair[0]-1, pair[0]+1, 13)
         z1_list = np.linspace(pair[1]-1, pair[1]+1, 13)
         z_list = []
@@ -137,7 +109,7 @@ def zstack_metadata_fixture(splitter_tmp_dir_fixture,
     return result
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def zstack_fixture(zstack_metadata_fixture,
                    z_to_exp_id_fixture,
                    splitter_tmp_dir_fixture):
@@ -173,42 +145,65 @@ def zstack_fixture(zstack_metadata_fixture,
 
         tifffile.imsave(tmp_dir / tiff_path, tiff_data)
 
-    return exp_id_to_expected
+    yield exp_id_to_expected
+
+    for exp_id in exp_id_to_expected:
+        this_path = pathlib.Path(exp_id_to_expected[exp_id])
+        if this_path.is_file():
+            this_path.unlink()
 
 
-@pytest.fixture(scope='session')
-def surface_fixture(splitter_tmp_dir_fixture):
+@pytest.fixture
+def surface_fixture(splitter_tmp_dir_fixture,
+                    roi_index_to_z_fixture):
     """
     Returns path to surface tiff and path to expected tiffs
     for each ROI
     """
+    n_rois = len(roi_index_to_z_fixture)
     tmp_dir = splitter_tmp_dir_fixture
     raw_tiff_path = tempfile.mkstemp(dir=tmp_dir,
                                      suffix='_surface.tiff')[1]
-    expected_0_path = tmp_dir / 'expected_surface_0.tiff'
-    expected_1_path = tmp_dir / 'expected_surface_1.tiff'
+    expected_path_list = []
+    for ii in range(n_rois):
+        expected_path = tmp_dir / f'expected_surface_{ii}.tiff'
+        expected_path_list.append(expected_path)
     n_pages = 7
 
+    data_list = []
+    expected_img_list = []
+
     rng = np.random.default_rng()
-    data_0 = rng.integers(0, 100, (n_pages, 24, 24))
-    expected_0 = normalize_array(array=data_0.mean(axis=0))
-    data_1 = rng.integers(100, 150, (n_pages, 24, 24))
-    expected_1 = normalize_array(array=data_1.mean(axis=0))
+    for ii in range(n_rois):
+        this_data = rng.integers(0, 100, (n_pages, 24, 24))
+        this_expected = normalize_array(this_data.mean(axis=0))
+        data_list.append(this_data)
+        expected_img_list.append(this_expected)
 
     tiff_data = []
     for i_page in range(n_pages):
-        tiff_data.append(data_0[i_page, :, :])
-        tiff_data.append(data_1[i_page, :, :])
+        for ii in range(n_rois):
+            tiff_data.append(data_list[ii][i_page, :, :])
+
     tifffile.imsave(raw_tiff_path, tiff_data)
-    tifffile.imsave(expected_0_path, expected_0)
-    tifffile.imsave(expected_1_path, expected_1)
+    for expected_path, expected_img in zip(expected_path_list,
+                                           expected_img_list):
+        tifffile.imsave(expected_path, expected_img)
 
-    return {'raw': raw_tiff_path,
-            'expected_0': str(expected_0_path.resolve().absolute()),
-            'expected_1': str(expected_1_path.resolve().absolute())}
+    result = dict()
+    result['raw'] = raw_tiff_path
+    for ii in range(n_rois):
+        str_path = str(expected_path_list[ii].resolve().absolute())
+        result[f'expected_{ii}'] = str_path
+    yield result
+
+    for key in result:
+        this_path = pathlib.Path(result[key])
+        if this_path.is_file():
+            this_path.unlink()
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def timeseries_fixture(splitter_tmp_dir_fixture,
                        image_metadata_fixture,
                        z_to_exp_id_fixture):
@@ -247,10 +242,15 @@ def timeseries_fixture(splitter_tmp_dir_fixture,
                 out_file.create_dataset('data', data=z_to_data[z_pair][zz])
             str_path = str(expected_path.resolve().absolute())
             result[f'expected_{exp_id}'] = str_path
-    return result
+    yield result
+
+    for key in result:
+        this_path = pathlib.Path(result[key])
+        if this_path.is_file():
+            this_path.unlink()
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def depth_fixture(splitter_tmp_dir_fixture,
                   image_metadata_fixture,
                   z_to_exp_id_fixture):
@@ -289,10 +289,15 @@ def depth_fixture(splitter_tmp_dir_fixture,
             tifffile.imsave(expected_path, expected_data)
             str_path = str(expected_path.resolve().absolute())
             result[f'expected_{exp_id}'] = str_path
-    return result
+    yield result
+
+    for key in result:
+        this_path = pathlib.Path(result[key])
+        if this_path.is_file():
+            this_path.unlink()
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture
 def input_json_fixture(
         depth_fixture,
         surface_fixture,
@@ -345,86 +350,3 @@ def input_json_fixture(
     params['plane_groups'] = plane_groups
     params['log_level'] = 'WARNING'
     return params
-
-
-def test_splitter_cli(input_json_fixture,
-                      tmp_path_factory,
-                      zstack_metadata_fixture,
-                      surface_metadata_fixture,
-                      image_metadata_fixture,
-                      depth_fixture,
-                      surface_fixture,
-                      timeseries_fixture,
-                      zstack_fixture):
-
-    tmp_dir = tmp_path_factory.mktemp('cli_output_json')
-    output_json = pathlib.Path(tmp_dir) / 'output.json'
-    output_json = str(output_json.resolve().absolute())
-    input_json = copy.deepcopy(input_json_fixture)
-    input_json['output_json'] = output_json
-
-    metadata_lookup = copy.deepcopy(zstack_metadata_fixture)
-    metadata_lookup[input_json['timeseries_tif']] = image_metadata_fixture
-    metadata_lookup[input_json['depths_tif']] = image_metadata_fixture
-    metadata_lookup[input_json['surface_tif']] = surface_metadata_fixture
-
-    def mock_read_metadata(tiff_path):
-        str_path = str(tiff_path.resolve().absolute())
-        return metadata_lookup[str_path]
-
-    to_replace = 'ophys_etl.modules.mesoscope_splitting.'
-    to_replace += 'tiff_metadata._read_metadata'
-    with patch(to_replace,
-               new=mock_read_metadata):
-
-        runner = TiffSplitterCLI(
-                    args=[],
-                    input_data=input_json)
-        runner.run()
-
-    exp_ct = 0
-    for plane_group in input_json['plane_groups']:
-        for exp in plane_group['ophys_experiments']:
-            exp_ct += 1
-            exp_dir = exp['storage_directory']
-            exp_dir = pathlib.Path(exp_dir)
-            exp_id = exp['experiment_id']
-            ts_actual = exp_dir / f'{exp_id}.h5'
-            with h5py.File(ts_actual, 'r') as in_file:
-                actual = in_file['data'][()]
-            ts_expected = timeseries_fixture[f'expected_{exp_id}']
-            with h5py.File(ts_expected, 'r') as in_file:
-                expected = in_file['data'][()]
-            np.testing.assert_array_equal(actual, expected)
-
-            depth_actual = exp_dir / f'{exp_id}_depth.tif'
-            with tifffile.TiffFile(depth_actual, 'rb') as in_file:
-                assert len(in_file.pages) == 1
-                actual = in_file.pages[0].asarray()
-            depth_expected = depth_fixture[f'expected_{exp_id}']
-            with tifffile.TiffFile(depth_expected, 'rb') as in_file:
-                assert len(in_file.pages) == 1
-                expected = in_file.pages[0].asarray()
-            np.testing.assert_array_equal(actual, expected)
-
-            surface_actual = exp_dir / f'{exp_id}_surface.tif'
-            with tifffile.TiffFile(surface_actual, 'rb') as in_file:
-                assert len(in_file.pages) == 1
-                actual = in_file.pages[0].asarray()
-
-            roi_index = exp['roi_index']
-            surface_expected = surface_fixture[f'expected_{roi_index}']
-            with tifffile.TiffFile(surface_expected, 'rb') as in_file:
-                assert len(in_file.pages) == 1
-                expected = in_file.pages[0].asarray()
-            np.testing.assert_array_equal(actual, expected)
-
-            stack_actual = exp_dir / f'{exp_id}_z_stack_local.h5'
-            with h5py.File(stack_actual, 'r') as in_file:
-                actual = in_file['data'][()]
-            stack_expected = zstack_fixture[f'expected_{exp_id}']
-            with h5py.File(stack_expected, 'r') as in_file:
-                expected = in_file['data'][()]
-            np.testing.assert_array_equal(actual, expected)
-
-    assert exp_ct == 8
